+++ src/sdk/index.js
@@ -83,6 +83,9 @@ export class qiscusSDK extends EventEmitter {
           vStore.dispatch('chatGroup', {id:QiscusSDK.core.selected.id, oldSelected})
         })
       },
+      getOrCreateRoomByChannel(channel, name, avatar_url) {
+        self.UI.getOrCreateRoomByUniqueId(channel, name, avatar_url);
+      },
       toggleChatWindow () {
         vStore.dispatch('toggleChatWindow')
       },
@@ -452,6 +455,10 @@ export class qiscusSDK extends EventEmitter {
       })
   }
 
+  getOrCreateRoomByChannel(channel, name, avatar_url) {
+    this.getOrCreateRoomByUniqueId(channel, name, avatar_url);
+  }
+
   /**
    * Set read status for selected comment
    * 
@@ -547,14 +554,16 @@ export class qiscusSDK extends EventEmitter {
   }
 
   loadComments (topic_id, last_comment_id = 0) {
-    return this.topicAdapter.loadComments(topic_id, last_comment_id)
-    .then((response) => {
-      this.selected.receiveComments(reverse(response))
-      this.sortComments()
-      return new Promise((resolve, reject) => resolve(response))
-    }, (error) => {
-      console.error('Error loading comments', error)
-    })
+    const self = this;
+    return self.topicAdapter.loadComments(topic_id, last_comment_id)
+      .then((response) => {
+        self.selected.receiveComments(response.reverse())
+        self.sortComments()
+        return new Promise((resolve, reject) => resolve(response))
+      }, (error) => {
+        console.error('Error loading comments', error)
+        return new Promise(reject => reject(error));
+      });
   }
 
   /**