+++ src/sdk/utils.js
@@ -66,6 +66,7 @@ export function scrollToBottom(latestCommentId) {
   requestAnimationFrame(function(){
     if(latestCommentId > 0){
       const elementToScroll = document.getElementById(latestCommentId)
+      if(!elementToScroll) return false;
       elementToScroll.scrollIntoView({block: 'end', behavior: 'smooth'})
     }
     //on entering the room, wait for data processed then focus on comment form